fixed(liis_sub): random_25-1 CORRECT. check if cur cover prev. fixed the liis_sub_check. random_25-2 and 3 still not correct.

+++ include/interval/sub.hpp
@@ -32,7 +32,7 @@ bool operator>(const sub &lhe, const sub &rhs);
 
 int miis_sub_algo(const std::deque<interval> &interval_seq);
 bool miis_sub_check(const std::deque<interval> &interval_seq, const int &q);
-bool miis_sub_comb(const std::deque<interval> &interval_seq, const int &q);
+bool miis_sub_comb(const std::deque<interval> &interval_seq, const size_t &q);
 
 int liis_sub_algo(const std::deque<interval> &interval_seq);
 bool liis_sub_check(const std::deque<interval> &interval_seq, const size_t &l);
